requirements.txt
requests
python-ldap
dnspython
